refactor(domain): use decimal for monetary values

+++ server/src/Application/Shopping/AddItemsToList.cs
@@ -12,7 +12,7 @@ public static class AddItemsToList {
   public record CommandItem(
     string? Name,
     string? Quantity,
-    float Price,
+    decimal Price,
     bool IsChecked
   );
