+++ server/src/Database/Entities/PurchaseDbEntity.cs
@@ -6,7 +6,7 @@ internal class PurchaseDbEntity {
   public int RegisteredItemId { get; set; }
   public int? ProductId { get; set; }
   public int? SuperMarketId { get; set; }
-  public float PricePaid { get; set; }
+  public decimal PricePaid { get; set; }
   public string? Quantity { get; set; }
   public DateTime PurchasedAt { get; set; }
