+++ server/test/UnitTests/Domain.UnitTests/Shopping/PriceTest.cs
@@ -5,24 +5,24 @@ namespace Metaspesa.Domain.UnitTests.Shopping;
 public static class PriceTest {
   public class IsZero {
     [Theory(DisplayName = "Returns true when value is below epsilon")]
-    [InlineData(0f)]
-    [InlineData(0.005f)]
-    [InlineData(0.009f)]
-    public void Price_IsZero_WhenValueBelowEpsilon(float value) {
+    [InlineData(0)]
+    [InlineData(0.005)]
+    [InlineData(0.009)]
+    public void Price_IsZero_WhenValueBelowEpsilon(double value) {
       // Arrange
-      var price = new Price(value);
+      var price = new Price((decimal)value);
 
       // Act & Assert
       Assert.True(price.IsZero());
     }
 
     [Theory(DisplayName = "Returns false when value is at or above epsilon")]
-    [InlineData(0.01f)]
-    [InlineData(0.5f)]
-    [InlineData(10f)]
-    public void Price_IsNotZero_WhenValueAtOrAboveEpsilon(float value) {
+    [InlineData(0.01)]
+    [InlineData(0.5)]
+    [InlineData(10)]
+    public void Price_IsNotZero_WhenValueAtOrAboveEpsilon(double value) {
       // Arrange
-      var price = new Price(value);
+      var price = new Price((decimal)value);
 
       // Act & Assert
       Assert.False(price.IsZero());
@@ -33,20 +33,20 @@ public static class PriceTest {
     [Fact(DisplayName = "Two prices with the same value are equal")]
     public void Price_Equal_WhenSameValue() {
       // Arrange
-      var price1 = new Price(5.0f);
-      var price2 = new Price(5.0f);
+      var price1 = new Price(5.0m);
+      var price2 = new Price(5.0m);
 
       // Act & Assert
       Assert.Equal(price1, price2);
     }
 
     [Theory(DisplayName = "Two prices within epsilon are equal")]
-    [InlineData(5.0f, 5.005f)]
-    [InlineData(5.0f, 4.995f)]
-    public void Price_Equal_WhenDifferenceWithinEpsilon(float a, float b) {
+    [InlineData(5.0, 5.005)]
+    [InlineData(5.0, 4.995)]
+    public void Price_Equal_WhenDifferenceWithinEpsilon(double a, double b) {
       // Arrange
-      var price1 = new Price(a);
-      var price2 = new Price(b);
+      var price1 = new Price((decimal)a);
+      var price2 = new Price((decimal)b);
 
       // Act & Assert
       Assert.Equal(price1, price2);
@@ -55,8 +55,8 @@ public static class PriceTest {
     [Fact(DisplayName = "Two prices with different values are not equal")]
     public void Price_NotEqual_WhenDifferentValues() {
       // Arrange
-      var price1 = new Price(5.0f);
-      var price2 = new Price(10.0f);
+      var price1 = new Price(5.0m);
+      var price2 = new Price(10.0m);
 
       // Act & Assert
       Assert.NotEqual(price1, price2);
@@ -65,8 +65,8 @@ public static class PriceTest {
     [Fact(DisplayName = "Equal prices have the same hash code")]
     public void Price_SameHashCode_WhenEqual() {
       // Arrange
-      var price1 = new Price(5.0f);
-      var price2 = new Price(5.0f);
+      var price1 = new Price(5.0m);
+      var price2 = new Price(5.0m);
 
       // Act & Assert
       Assert.Equal(price1.GetHashCode(), price2.GetHashCode());
